# perf, ringbuf: add ReadBuffer mehods