perf: add Reader.ReadBuffer

    Allow reading from a Reader without allocations, once a steady state
    has been reached.

        name      old time/op    new time/op    delta
        Reader-4    8.98µs ± 2%    8.91µs ± 3%     ~     (p=0.686 n=4+4)

        name      old alloc/op   new alloc/op   delta
        Reader-4      424B ± 0%      288B ± 0%  -32.08%  (p=0.029 n=4+4)

        name      old allocs/op  new allocs/op  delta
        Reader-4      8.00 ± 0%      1.00 ± 0%  -87.50%  (p=0.029 n=4+4)

ringbuf: fix BenchmarkReader

    Move calls to ResetTimer and ReportAllocs and avoid allocating a
    temporary buffer.

ringbuf: add ReadBuffer method

    Add a ReadBuffer method to ringbuf.Reader, analogous to perf.Reader.

        name                                   old time/op    new time/op    delta
        Reader/normal_epoll_with_timeout_-1-4    8.58µs ± 3%    9.07µs ± 4%     ~     (p=0.057 n=4+4)

        name                                   old alloc/op   new alloc/op   delta
        Reader/normal_epoll_with_timeout_-1-4      384B ± 0%      288B ± 0%  -25.00%  (p=0.029 n=4+4)

        name                                   old allocs/op  new allocs/op  delta
        Reader/normal_epoll_with_timeout_-1-4      4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.029 n=4+4)