@lmb 2022-05-10
I've changed the signature to `ReadRecord(*Record) error` based on feedback from Timo. Maybe `ReadInto(*Record)` would be better?